Print timestamp as an unsigned long, which works a little better 	on 64-bit systems.
	Patch from Andrew Belashov

+++ src/main/xlat.c
@@ -544,7 +544,8 @@ int radius_xlat(char *out, int outlen, const char *fmt,
 				p++;
 				break;
 			case 'l': /* request timestamp */
-				snprintf(tmpdt, sizeof(tmpdt), "%lu",request->timestamp);
+				snprintf(tmpdt, sizeof(tmpdt), "%lu",
+					 (unsigned long) request->timestamp);
 				strNcpy(q,tmpdt,freespace);
 				q += strlen(q);
 				p++;
